+++ app/src/main/java/com/example/composeweather/domain/model/Hourly.kt
@@ -3,7 +3,7 @@ package com.example.composeweather.domain.model
 data class Hourly(
     val clouds: Int,
     val dew_point: Double,
-    val dt: Int,
+    val dt: Double,
     val feels_like: Double,
     val humidity: Int,
     val pop: Double,
